+++ src/main/java/com/idt/aio/repository/DocumentKeywordRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentKeywordRepository extends JpaRepository<DocumentKeyword, Long> {
+public interface DocumentKeywordRepository extends JpaRepository<DocumentKeyword, Integer> {
 }
