+++ src/main/java/com/idt/aio/repository/HomonymRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HomonymRepository extends JpaRepository<Homonym, Long> {
+public interface HomonymRepository extends JpaRepository<Homonym, Integer> {
 }
