+++ src/main/java/com/idt/aio/repository/ModelPresetRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ModelPresetRepository extends JpaRepository<ModelPreset, Long> {
+public interface ModelPresetRepository extends JpaRepository<ModelPreset, Integer> {
 }
